+++ CHANGELOG.md
@@ -47,6 +47,8 @@ to docs, or any other relevant information.
   `LoadClientConfigOptions::builder()` to construct client configuration loading options.
 * `LoadClientConfigProfileOptions` is now tagged with `#[non_exhaustive]`. Use
   `LoadClientConfigProfileOptions::builder()` to construct profile loading options.
+* `ClientConfigFromTOMLOptions` is now tagged with `#[non_exhaustive]`. Use
+  `ClientConfigFromTOMLOptions::builder()` to construct TOML parsing options.
 
 ## [0.6.0] - 2026-08-04